Multiple changes: - Fixed a bug in BoundingSphericalSector that caused CDLOD to scale linearry with the area of the heightmap. - Also added statictics to GridMesh

+++ src/cpp/engine/cdlod/quad_tree_node.h
@@ -15,7 +15,7 @@ class QuadTreeNode {
  public:
   QuadTreeNode(int x, int z, GLubyte level, int dimension);
 
-  GLushort size() { return dimension_ << level_; }
+  int size() { return dimension_ << level_; }
 
   bool collidesWithSphere(const glm::vec3& center, float radius) {
     return bbox_.collidesWithSphere(center, radius);
